+++ titowebhooks/admin.py
@@ -13,7 +13,7 @@ def send_to_sendy_action(modeladmin, request, queryset):
         for event in queryset:
             try:
                 async_task(
-                    "titowebhooks.utils.send_to_sendy",
+                    "sendy.utils.send_to_sendy",
                     email=event.payload["email"],
                     name=f"{event.payload['first_name']} {event.payload['last_name']}",
                     campaign_id=sendy_list.list_id,
